requirements.txt
Django==2.2.1
django-extensions==2.1.6
django-filter==2.1.0
djangorestframework==3.9.4
Markdown==3.1
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
Unidecode==1.0.23
whitenoise==4.1.2
